+++ vec4state.h
@@ -18,7 +18,7 @@ class vec4state
         // Returns true if the vector contains x or z, otherwise returns false
         bool isUnknown() const;
         // Returns the number of VPI elements in the vector
-        int getVectorSize() const;
+        size_t getVectorSize() const;
 
     public:
         // C'tors
@@ -27,10 +27,10 @@ class vec4state
         // Initializes a 32-bit vector to the value of num
         vec4state(int num);
         // Initializes a 64-bit vector to the value of num
-        vec4state(long num);
+        vec4state(long long num);
         // Initializes a vector of size str.length with the value of str
         vec4state(string str);
-        // Initializes a vector of size "size" with the value of str
+        // Initializes a vector of size "size" that contains str (of length 1) repeated size times. 
         vec4state(string str, size_t size);
         // Copy c'tor
         vec4state(const vec4state& other);
@@ -41,23 +41,23 @@ class vec4state
         // Assignment operators
         vec4state& operator=(const vec4state& other);
         vec4state& operator=(int num);
-        vec4state& operator=(long num);
+        vec4state& operator=(long long num);
         vec4state& operator=(string str);
 
         // Bitwise operators
         vec4state operator&(const vec4state& other) const;
-        vec4state operator&(long num) const;
+        vec4state operator&(long long num) const;
         vec4state operator|(const vec4state& other) const;
-        vec4state operator|(long num) const;
+        vec4state operator|(long long num) const;
         vec4state operator^(const vec4state& other) const;
-        vec4state operator^(long num) const;
+        vec4state operator^(long long num) const;
         vec4state operator~() const;
 
         // Equality operators
         vec4state operator==(const vec4state& other) const;
-        vec4state operator==(long num) const;
+        vec4state operator==(long long num) const;
         vec4state operator!=(const vec4state& other) const;
-        vec4state operator!=(long num) const;
+        vec4state operator!=(long long num) const;
         // Checks if the vector is equal to other, including x's and z's. Returns a vector of 1's if true and 0's otherwise.
         vec4state caseEquality(const vec4state& other) const;
         // Checks if the vector is not equal to other, including x's and z's. Returns a vector of 1's if true and 0's otherwise.
@@ -66,60 +66,60 @@ class vec4state
         // Shift operators
         // If other contains x or z, returns a vector that is all x's.
         vec4state operator<<(const vec4state& other);
-        vec4state operator<<(const long num);
+        vec4state operator<<(const long long num);
         // If other contains x or z, returns a vector that is all x's.
         vec4state operator>>(const vec4state& other);
-        vec4state operator>>(const long num);
+        vec4state operator>>(const long long num);
 
         // Slice methods
         vec4state operator[](const vec4state& index);
-        vec4state operator[](const long index);
-        // Returns a vector that is a slice of the original vector from start to end
-        vec4state getSlice(long end, long start);
-        // Sets the slice of the original vector from start to end to the value of other
-        void setSlice(long end, long start, const vec4state& other);
+        vec4state operator[](const long long index);
+        // Returns a vector that is a slice of the original vector from start to end.
+        vec4state getSlice(long long end, long long start);
+        // Sets the slice of the original vector from start to end to the value of other.
+        void setSlice(long long end, long long start, const vec4state& other);
 
         // Logical operators
         vec4state operator&&(const vec4state& other) const;
-        vec4state operator&&(long num) const;
+        vec4state operator&&(long long num) const;
         vec4state operator||(const vec4state& other) const;
-        vec4state operator||(long num) const;
+        vec4state operator||(long long num) const;
         vec4state operator!() const;
 
         // Relational operators
         vec4state operator<(const vec4state& other) const;
-        vec4state operator<(long num) const;
+        vec4state operator<(long long num) const;
         vec4state operator>(const vec4state& other) const;
-        vec4state operator>(long num) const;
+        vec4state operator>(long long num) const;
         vec4state operator<=(const vec4state& other) const;
-        vec4state operator<=(long num) const;
+        vec4state operator<=(long long num) const;
         vec4state operator>=(const vec4state& other) const;
-        vec4state operator>=(long num) const;
+        vec4state operator>=(long long num) const;
 
         // Arithmetic operators
         vec4state operator+(const vec4state& other) const;
-        vec4state operator+(long num) const;
+        vec4state operator+(long long num) const;
         vec4state operator-(const vec4state& other) const;
-        vec4state operator-(long num) const;
+        vec4state operator-(long long num) const;
         vec4state operator*(const vec4state& other) const;
-        vec4state operator*(long num) const;
+        vec4state operator*(long long num) const;
         vec4state operator/(const vec4state& other) const;
-        vec4state operator/(long num) const;
+        vec4state operator/(long long num) const;
         vec4state operator%(const vec4state& other) const;
-        vec4state operator%(long num) const;
+        vec4state operator%(long long num) const;
         vec4state operator-() const;
-        // Calculates the value of the vector to the power of other
+        // Calculates the value of the vector to the power of other.
         vec4state power(const vec4state& other) const;
-        // Calculates the value of the vector to the power of num
-        vec4state power(long num) const;
+        // Calculates the value of the vector to the power of num.
+        vec4state power(long long num) const;
         
         // Casting operators
         // TODO: Figure out how to implement these
 
-        // Getters
-        // Returns the size of the vector
+        /*********** For testing purposes ***********/
+        // Returns the size of the vector.
         size_t getSize() const;
-        // Returns the value of the vector as a string
+        // Returns a string representation of the vector.
         string toString() const;
 };
